split the file into modules
